# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Learn Kids API
  description: API платформы обучения робототехнике для семьи
  version: 1.0.0

servers:
  - url: /api/v1
    description: API v1

paths:
  /health:
    get:
      summary: Проверка готовности сервера и БД
      tags: [health]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '503':
          description: Сервис недоступен (БД не отвечает)

  /auth/register:
    post:
      summary: Регистрация пользователя
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password, name]
              properties:
                login: { type: string }
                password: { type: string }
                name: { type: string }
                role: { type: string, enum: [student, teacher], default: student }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '400':
          description: Неверный запрос

  /auth/login:
    post:
      summary: Вход
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '401':
          description: Неверный логин или пароль

  /auth/me:
    get:
      summary: Текущий пользователь
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: Не авторизован
    patch:
      summary: Обновить профиль (тема)
      tags: [auth]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                theme: { type: string, enum: [default, light, cyberpunk] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Неверное тело запроса
        '401':
          description: Не авторизован

  /modules:
    get:
      summary: Список модулей
      tags: [lessons]
      parameters:
        - name: platform
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Module' }
    post:
      summary: Создать курс (только teacher)
      tags: [lessons]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                sort_order: { type: integer }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '400':
          description: Неверное тело запроса
        '401':
          description: Не авторизован
        '403':
          description: Только для учителя

  /modules/{id}:
    delete:
      summary: Удалить курс (только teacher)
      tags: [lessons]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Удалено
        '401':
          description: Не авторизован
        '403':
          description: Только для учителя
        '404':
          description: Модуль не найден
    get:
      summary: Модуль по ID с вложенными уроками
      tags: [lessons]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '404':
          description: Модуль не найден

  /modules/{id}/lessons:
    post:
      summary: Добавить урок в курс (только teacher)
      tags: [lessons]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                lesson_type: { type: string, enum: [theory, practice, project], default: theory }
                sort_order: { type: integer }
                steps:
                  type: array
                  items:
                    type: object
                    required: [title]
                    properties:
                      title: { type: string }
                      content: { type: string }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lesson' }
        '400':
          description: Неверное тело запроса
        '401':
          description: Не авторизован
        '403':
          description: Только для учителя
        '404':
          description: Модуль не найден

  /lessons/{id}/comments:
    get:
      summary: Список комментариев к уроку
      tags: [comments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LessonComment' }
        '400':
          description: Неверный id урока
    post:
      summary: Добавить комментарий (требуется авторизация)
      tags: [comments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LessonComment' }
        '400':
          description: Неверное тело запроса
        '401':
          description: Не авторизован

  /lessons/{id}/reaction:
    put:
      summary: Поставить лайк/дизлайк уроку
      tags: [reactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reaction]
              properties:
                reaction: { type: string, enum: [like, dislike] }
      responses:
        '204':
          description: OK
        '400':
          description: Неверный id или reaction
        '401':
          description: Не авторизован
    delete:
      summary: Убрать свою реакцию к уроку
      tags: [reactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: OK
        '401':
          description: Не авторизован

  /lessons/{id}/comments/{commentId}:
    delete:
      summary: Удалить свой комментарий
      tags: [comments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: ID урока
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: ID комментария
      responses:
        '204':
          description: Удалено
        '401':
          description: Не авторизован
        '404':
          description: Комментарий не найден или доступ запрещён

  /lessons/{id}/comments/{commentId}/reaction:
    put:
      summary: Поставить лайк/дизлайк комментарию
      tags: [reactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: ID урока
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: ID комментария
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reaction]
              properties:
                reaction: { type: string, enum: [like, dislike] }
      responses:
        '204':
          description: OK
        '400':
          description: Неверный id или reaction
        '401':
          description: Не авторизован
    delete:
      summary: Убрать свою реакцию к комментарию
      tags: [reactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: OK
        '401':
          description: Не авторизован

  /lessons/{id}:
    get:
      summary: Урок по ID (шаги, материалы, чек-лист)
      tags: [lessons]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lesson' }
        '404':
          description: Урок не найден
    delete:
      summary: Удалить урок (только teacher)
      tags: [lessons]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Удалено
        '401':
          description: Не авторизован
        '403':
          description: Только для учителя
        '404':
          description: Урок не найден
    put:
      summary: Обновить урок (только teacher)
      tags: [lessons]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                steps:
                  type: array
                  items:
                    type: object
                    required: [title]
                    properties:
                      title: { type: string }
                      content: { type: string }
                      sort_order: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lesson' }
        '400':
          description: Неверное тело запроса
        '401':
          description: Не авторизован
        '403':
          description: Только для учителя
        '404':
          description: Урок не найден

  /progress:
    get:
      summary: Прогресс текущего пользователя
      tags: [progress]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProgress' }
        '401':
          description: Не авторизован

  /lessons/{id}/progress:
    put:
      summary: Установить статус по уроку
      tags: [progress]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [not_started, in_progress, completed]
      responses:
        '200':
          description: OK
        '401':
          description: Не авторизован

  /lessons/{id}/checklist/{itemId}:
    put:
      summary: Отметить пункт чек-листа
      tags: [progress]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                completed: { type: boolean }
      responses:
        '200':
          description: OK
        '401':
          description: Не авторизован

  /users:
    get:
      summary: Список пользователей (только teacher)
      tags: [parent-dashboard]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '403':
          description: Только для учителя

  /users/{id}/progress:
    get:
      summary: Прогресс ученика (только teacher)
      tags: [parent-dashboard]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProgress' }
        '403':
          description: Только для учителя

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        login: { type: string }
        name: { type: string }
        role: { type: string, enum: [student, teacher] }
        theme: { type: string, enum: [default, light, cyberpunk], description: Цветовая схема портала }
        created_at: { type: string, format: date-time }
    Module:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        sort_order: { type: integer }
        lessons: { type: array, items: { $ref: '#/components/schemas/Lesson' } }
        created_at: { type: string, format: date-time }
    Lesson:
      type: object
      properties:
        id: { type: string, format: uuid }
        module_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        lesson_type: { type: string, enum: [theory, practice, project] }
        sort_order: { type: integer }
        steps: { type: array, items: { $ref: '#/components/schemas/LessonStep' } }
        materials: { type: array, items: { $ref: '#/components/schemas/LessonMaterial' } }
        tags: { type: array, items: { type: string } }
        checklist: { type: array, items: { $ref: '#/components/schemas/ChecklistItem' } }
        created_at: { type: string, format: date-time }
        likes_count: { type: integer, description: "Количество лайков" }
        dislikes_count: { type: integer, description: "Количество дизлайков" }
        user_reaction: { type: string, enum: [like, dislike], description: "Реакция текущего пользователя (если авторизован)" }
    LessonStep:
      type: object
      properties:
        id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        title: { type: string }
        content: { type: string }
        sort_order: { type: integer }
    LessonMaterial:
      type: object
      properties:
        id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        kind: { type: string, enum: [link, file] }
        url_or_path: { type: string }
        title: { type: string }
    ChecklistItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        title: { type: string }
        sort_order: { type: integer }
    LessonComment:
      type: object
      properties:
        id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        user_name: { type: string }
        text: { type: string }
        created_at: { type: string, format: date-time }
        likes_count: { type: integer, description: "Количество лайков" }
        dislikes_count: { type: integer, description: "Количество дизлайков" }
        user_reaction: { type: string, enum: [like, dislike], description: "Реакция текущего пользователя (если авторизован)" }
    UserProgress:
      type: object
      properties:
        lessons:
          type: array
          items:
            type: object
            properties:
              lesson_id: { type: string, format: uuid }
              status: { type: string, enum: [not_started, in_progress, completed] }
              updated_at: { type: string, format: date-time }
        checklist:
          type: array
          items:
            type: object
            properties:
              checklist_item_id: { type: string, format: uuid }
              completed_at: { type: string, format: date-time, nullable: true }