# backend/internal/server/server.go
package server

import (
	"context"
	"net/http"
	"os"
	"strings"
	"time"

	"learn_kids/backend/api"
	"learn_kids/backend/internal/domain/chat"
	"learn_kids/backend/internal/domain/comments"
	"learn_kids/backend/internal/domain/lessons"
	"learn_kids/backend/internal/domain/progress"
	"learn_kids/backend/internal/domain/reactions"
	"learn_kids/backend/internal/domain/users"
	"learn_kids/backend/internal/httplog"
	"learn_kids/backend/internal/middleware"

	"github.com/gin-gonic/gin"
	"github.com/jackc/pgx/v5/pgxpool"
)

type Server struct {
	engine   *gin.Engine
	pool     *pgxpool.Pool
	lessons  *lessons.Handler
	users    *users.Handler
	progress *progress.Handler
	chat      *chat.Handler
	comments  *comments.Handler
	reactions *reactions.Handler

	// shutdown context: when cancelled, rate limiter cleanup goroutines exit
	shutdownContext context.Context

	// cached middleware closures (avoid re-creation on every request)
	requireAuthMw    gin.HandlerFunc
	requireTeacherMw gin.HandlerFunc
}

type Deps struct {
	Pool             *pgxpool.Pool
	Lessons          *lessons.Handler
	Users            *users.Handler
	Progress         *progress.Handler
	Chat             *chat.Handler
	Comments         *comments.Handler
	Reactions        *reactions.Handler
	JWTSecret        string
	FrontendOrigin   string
	ShutdownContext  context.Context // optional: cancels on server shutdown so rate limiters stop cleanup
}

func New(d Deps) *Server {
	gin.SetMode(gin.ReleaseMode)
	engine := gin.New()

	// Only trust loopback and private Docker/LAN subnets as reverse proxies.
	// This ensures c.ClientIP() returns the real client IP (not spoofed X-Forwarded-For).
	_ = engine.SetTrustedProxies([]string{"127.0.0.1", "::1", "10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"})

	origin := d.FrontendOrigin
	if origin == "" {
		origin = "*"
	}
	engine.Use(gin.Recovery())
	engine.Use(middleware.CORS(origin))
	engine.Use(middleware.RequestID())
	engine.Use(middleware.HTTPLog())

	// Max request body size: 1 MB (protection against OOM)
	engine.Use(func(c *gin.Context) {
		c.Request.Body = http.MaxBytesReader(c.Writer, c.Request.Body, 1<<20)
		c.Next()
	})

	s := &Server{
		engine:           engine,
		pool:             d.Pool,
		lessons:          d.Lessons,
		users:            d.Users,
		progress:         d.Progress,
		chat:             d.Chat,
		comments:         d.Comments,
		reactions:        d.Reactions,
		shutdownContext:  d.ShutdownContext,
		requireAuthMw:    middleware.RequireAuth(),
		requireTeacherMw: middleware.RequireTeacher(),
	}
	s.routes()
	return s
}

func (s *Server) routes() {
	s.engine.GET("/api/v1/health", s.health)
	s.engine.GET("/api/docs", func(c *gin.Context) {
		b, _ := api.FS.ReadFile("openapi.yaml")
		c.Data(http.StatusOK, "application/x-yaml", b)
	})

	// Rate limiters (shutdown context stops cleanup goroutines on server exit)
	authLimiter := middleware.NewRateLimiter(10, time.Minute, s.shutdownContext)
	chatLimiter := middleware.NewRateLimiter(20, time.Minute, s.shutdownContext)

	api := s.engine.Group("/api/v1")
	api.Use(middleware.Auth(s.users))

	// Public auth (rate limited)
	api.POST("/auth/register", authLimiter.Handler(), s.users.RegisterUser)
	api.POST("/auth/login", authLimiter.Handler(), s.users.Login)

	// Lessons (public read; teacher can create/update)
	api.GET("/modules", s.lessons.ListModules)
	api.GET("/modules/:id", s.lessons.GetModule)
	api.GET("/lessons/:id", s.lessons.GetLesson)
	api.PUT("/lessons/:id", s.requireAuth, s.requireTeacher, s.lessons.UpdateLesson)
	api.DELETE("/lessons/:id", s.requireAuth, s.requireTeacher, s.lessons.DeleteLesson)
	api.POST("/modules", s.requireAuth, s.requireTeacher, s.lessons.CreateModule)
	api.DELETE("/modules/:id", s.requireAuth, s.requireTeacher, s.lessons.DeleteModule)
	api.POST("/modules/:id/lessons", s.requireAuth, s.requireTeacher, s.lessons.CreateLesson)

	// Comments (public read; auth required to post)
	api.GET("/lessons/:id/comments", s.comments.List)
	api.POST("/lessons/:id/comments", s.requireAuth, s.comments.Create)
	api.DELETE("/lessons/:id/comments/:commentId", s.requireAuth, s.comments.Delete)

	// Reactions (likes/dislikes) for lessons and comments — JWT required to set/delete
	api.PUT("/lessons/:id/reaction", s.requireAuth, s.reactions.SetLessonReaction)
	api.DELETE("/lessons/:id/reaction", s.requireAuth, s.reactions.DeleteLessonReaction)
	api.PUT("/lessons/:id/comments/:commentId/reaction", s.requireAuth, s.reactions.SetCommentReaction)
	api.DELETE("/lessons/:id/comments/:commentId/reaction", s.requireAuth, s.reactions.DeleteCommentReaction)

	// Auth required from here for some routes
	api.GET("/auth/me", s.requireAuth, s.users.Me)
	api.PATCH("/auth/me", s.requireAuth, s.users.UpdateMe)
	api.GET("/progress", s.requireAuth, s.progress.GetProgress)
	api.PUT("/lessons/:id/progress", s.requireAuth, s.progress.SetLessonProgress)
	api.PUT("/lessons/:id/checklist/:itemId", s.requireAuth, s.progress.SetChecklistItem)

	// Chat with Gemini (auth required, rate limited)
	api.POST("/chat", s.requireAuth, chatLimiter.Handler(), s.chat.Chat)
	api.GET("/chat/:lessonId/history", s.requireAuth, s.chat.GetHistory)
	api.DELETE("/chat/:lessonId/history", s.requireAuth, s.chat.ClearHistory)

	// Teacher only
	api.GET("/users", s.requireAuth, s.requireTeacher, s.users.ListUsers)
	api.DELETE("/users/:id", s.requireAuth, s.requireTeacher, s.users.DeleteUser)
	api.GET("/users/:id/progress", s.requireAuth, s.requireTeacher, s.progress.GetUserProgress)

	// Frontend static (optional: serve from ./web when present, e.g. in Docker)
	if _, err := os.Stat("./web/index.html"); err == nil {
		s.engine.Static("/assets", "./web/assets")
		s.engine.StaticFile("/vite.svg", "./web/vite.svg")
		s.engine.NoRoute(func(c *gin.Context) {
			if strings.HasPrefix(c.Request.URL.Path, "/api/") {
				c.JSON(http.StatusNotFound, gin.H{"error": "not found"})
				return
			}
			c.File("./web/index.html")
		})
	}
}

func (s *Server) requireAuth(c *gin.Context) {
	s.requireAuthMw(c)
}

func (s *Server) requireTeacher(c *gin.Context) {
	s.requireTeacherMw(c)
}

func (s *Server) health(c *gin.Context) {
	if err := s.pool.Ping(c.Request.Context()); err != nil {
		httplog.LogError(c, err)
		c.JSON(http.StatusServiceUnavailable, gin.H{"status": "unhealthy"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"status": "ok"})
}

func (s *Server) Run(addr string) error {
	return s.engine.Run(addr)
}

func (s *Server) Handler() http.Handler {
	return s.engine
}